Partial tie-in of player.html. New blueprint section "players", handles get_rank and view_player now. Forms updated to use steam.py
