+++ functions.py
@@ -33,43 +33,6 @@ def get_replay_path(uid, add_extension=True):
     return os.path.join(replay_dir, uid + ('.gz' if add_extension else ''))
 
 
-rank_cache = {}
-
-
-def get_rank(steam_id):
-    if len(str(steam_id)) < 17:
-        return {}
-    if steam_id in rank_cache:
-        return rank_cache[steam_id]
-    url = "https://api.rocketleaguestats.com/v1/player?unique_id={}&platform_id=1".format(steam_id)
-    post_data = {'Authorization': config.RLSTATS_API_KEY}
-    data = requests.get(url, headers=post_data)
-    data = data.json()
-    # print (data)
-    if 'rankedSeasons' in data:
-        seasons = {}
-        for k in data['rankedSeasons']:
-            season = data['rankedSeasons'][k]
-            modes = []
-
-            names = {'13': 'standard', '11': 'doubles', '10': 'duel', '12': 'solo'}
-            for t in season:
-                if 'tier' in season[t]:  # excludes unranked
-                    s = {'mode': names[t], 'rank_points': season[t]['rankPoints'], 'tier': season[t]['tier'],
-                         'division': season[t]['division'],
-                         'string': tier_div_to_string(season[t]['tier'], season[t]['division'])}
-                    modes.append(s)
-                else:
-                    s = {'mode': names[t], 'rank_points': season[t]['rankPoints'], 'tier': 0,
-                         'division': 0, 'string': tier_div_to_string(0, 0)}
-                    modes.append(s)
-            seasons[k] = modes
-        rank_cache[steam_id] = seasons
-        return seasons
-    else:
-        return {}
-
-
 def tier_div_to_string(rank, div=-1):
     ranks = ['Unranked', 'Bronze I', 'Bronze II', 'Bronze III', 'Silver I', 'Silver II', 'Silver III', 'Gold I',
              'Gold II', 'Gold III', 'Platinum I', 'Platinum II', 'Platinum III', 'Diamond I', 'Diamond II',
