+++ static/index.css
@@ -12,6 +12,7 @@ html, body {
     background: white;
     padding: 20px;
     border: 1px solid black;
+    box-shadow: 0px 0px 10px 2px #cccccc;
     display: flex;
     margin: 12% auto auto auto;
     flex-direction: column;
