+++ steam.py
@@ -4,13 +4,14 @@ from config import STEAM_API_KEY
 
 bp = Blueprint('steam', __name__, url_prefix='/steam')
 
+
 @bp.route('/resolve', methods=['POST'])
 def resolve_steam():
     if 'name' not in request.form:
         return jsonify({})
     r = vanity_to_steam_id(request.form['name'])
     steamid = r['response']['steamid']
-    return redirect(url_for('view_player', id_=steamid))
+    return redirect(url_for('players.view_player', id_=steamid))
 
 
 def steam_id_to_profile(steamID):
