+++ templates/admin.html
@@ -1,4 +1,4 @@
-{% include 'header.html' %}
+{% include 'partials/header.html' %}
 <body>
 <div class="container">
     <h2>Submit new config file</h2>
@@ -43,4 +43,4 @@
     <div id="uploads-day"></div>
 </div>
 <script src="/static/admin.js"></script>
-{% include "footer.html" %}
+{% include "partials/footer.html" %}
