+++ templates/api.html
@@ -1,4 +1,4 @@
-{% include "header.html" %}
+{% include "partials/header.html" %}
 
 <div class="container">
     <h2>Saltie Replay API</h2>
@@ -104,4 +104,4 @@
     </table>
 </div>
 
-{% include "footer.html" %}
+{% include "partials/footer.html" %}
