+++ templates/error.html
@@ -1,4 +1,4 @@
-{% include 'header.html' %}
+{% include 'partials/header.html' %}
 <main role="main">
     <div class="container">
         <p>{{ error }}</p>
@@ -6,4 +6,4 @@
 </main>
 
 
-{% include 'footer.html' %}
+{% include 'partials/footer.html' %}
