+++ templates/replay.html
@@ -1,4 +1,4 @@
-{% include 'header.html' %}
+{% include 'partials/header.html' %}
 {% set frame_count = replay.frames|length %}
 <main role="main">
     <div class="container">
@@ -133,4 +133,4 @@
     </div>
 </main>
 
-{% include 'footer.html' %}
+{% include 'partials/footer.html' %}
