+++ templates/upload.html
@@ -1,4 +1,4 @@
-{% include 'header.html' %}
+{% include 'partials/header.html' %}
 
 <main role="main">
     <div class="container">
@@ -61,4 +61,4 @@
         </table>
     </div>
 </main>
-{% include "footer.html" %}
+{% include "partials/footer.html" %}
